Added Hex input option
This works in conjunction with the RGBA picker. When you change one, the
other changes with it. This works well to provide support for those who
want to use a GUI colour picker, with Hex inputs alongside, as well as
an easy way to preview Hex colours as they're inserted.